@@ -16,7 +16,7 @@ keywords = ["lexer", "lexical", "tokenizer", "parser", "no_std"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/maciejhirsz/logos"
-rust-version = "1.74.0"
+rust-version = "1.80.0"
 version = "0.15.1"
 include = ["build.rs", "/src/", "/examples/", "/tests/", "LICENSE-MIT", "LICENSE-APACHE"]
 
